FIX: 랜더링 오류 수정
